requirements.txt
idna==3.7
markupsafe==2.1.3
awkward-cpp==35
requests==2.32.2
setuptools==69.5.1
urllib3==2.2.2
bottleneck==1.3.7
sympy==1.12
awkward==2.6.6
fonttools==4.51.0
tornado==6.4.1
pillow==10.4.0
pyzmq==25.1.2
pyyaml==6.0.1
torchvision==0.18.1
wheel==0.43.0
brotli==1.0.9
gpytorch==1.12
torch==2.3.1
pyparsing==3.0.9
mkl-service==2.4.0
mkl-random==1.2.4
mpmath==1.3.0
pyqt5==5.15.10
linear-operator==0.5.2
joblib==1.4.2
debugpy==1.6.7
threadpoolctl==3.5.0
pytz==2024.1
contourpy==1.2.0
filelock==3.13.1
uproot==5.3.10
fsspec==2024.6.1
numpy==1.26.4
unicodedata2==15.1.0
jinja2==3.1.4
pyqt5-sip==12.13.0
networkx==3.3
matplotlib==3.8.4
torchaudio==2.3.1
pysocks==1.7.1
certifi==2024.7.4
gmpy2==2.1.2
psutil==5.9.0
typeguard==2.13.3
pandas==2.2.2
scikit-learn==1.5.1
mkl-fft==1.3.8
jupyter-core==5.7.2
pip==24.0
sip==6.7.12
ply==3.11
scipy==1.14.0
kiwisolver==1.4.4
numexpr==2.8.7
cramjam==2.8.3
jaxtyping==0.2.33
